Initial import of µTest
This is a project I've been working off an on for about a couple of
years in my spare time. After 10 years of using the GLib testing API
and hitting all its shortcomings, I was finally fed up and decided
to write my own C testing API.

µTest is heavily inspired by the Mocha and Jasmine testing frameworks
for JavaScript, to the point of borrowing their API and default output.
If the results are presented to the user console, you'll get fancy
colors and other effects; of course, redirecting to a non-TTY output
will drop anything that is not pure text.

Since I like TAP, µTest includes a TAP output formatter as well, so it's
possible to use the TAP harness in Autotools.

µTest is meant to be used as a sub-project in Meson, to minimise the
system dependencies.