Allow custom comparison functions
We require a comparison function passed to mutest_expect() to conform to
mutest_expect_func_t, and we provide a fair number of them out of the
box, but we should allow users of µTest to create their own comparators.
Since we can't know what kind of arguments these custom comparators
accept, we collect a pointer to a mutest_expect_res_t.