Rework matchers to allow generic ones
API break: renamed `mutest_expect_func_t` to `mutest_matcher_func_t`,
which is more accurate.

Alongside the exact type matchers, we want generic ones, that use the
type of the value passed to mutest_expect() as the type for the value
passed to the matcher function.

This means we can add matchers like:

 - mutest_to_be
 - mutest_to_be_in_range
 - mutest_to_be_close_to

On top of this work, we can now add matchers for floating point values
like NaN, +Inf, and -Inf, as well as matchers for order relations, like
less (or equal) than and greater (or equal) than.

The exact type matchers are still there, but on the chopping block.