Allow building µTest as a pure ancillary library
We want to allow building and installing µTest as a shared library by
default, but when used as a sub-project in another library, we may want
to allow building µTest as a pure static ancillary library that does not
get installed on the system.

In order to do that we add a new option, called `static`, and we check
if both the `static` option is set *and* if we're used as a sub-project.