Allow specifying a reason for an expectation skip
The `mutest_skip()` matcher should collect the reason why we're
skipping, so we can use it in output formatters like the TAP one.

Since `mutest_skip()` immediately breaks the `mutest_expect()` loop on
the varargs, and it should be the last matcher, we're either going to
collect a `char*` or we're going to collect the terminating NULL; either
way, we won't be going over the length of the variadic arguments list.