Add fallbacks for mutest_get_current_time()
We should have a fallback in place for platforms that do not support
clock_gettime() or the Windows performance counters.

Luckily, macOS has a simple API for a monotonic clock in their Mach
kernel API that we can use.

If all else fails, we can fall back to gettimeofday() which isn't
monotonic at all, but if you're dealing with short time intervals it can
be close enough to at least not be terrible.