tap: Add version preamble
Starting with version 13, the TAP specification mandates the version to
be the first thing printed on standard output. Conformant TAP harnesses
have started warning about a missing version.

Since we can have multiple suites per binary, we need to add a new
preamble hook to the formatter vtable, which gets called by
mutest_init(). This way, we can print out the TAP version header, and
then traverse the suites and specs as before.